+++ src/vec.rs
@@ -2,25 +2,25 @@ use std::ops;
 
 #[derive(Debug, Copy, Clone)]
 pub struct Vec3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl Vec3 {
-    pub fn new(x: f64, y: f64, z: f64) -> Vec3 {
+    pub fn new(x: f32, y: f32, z: f32) -> Vec3 {
         Vec3 { x, y, z }
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self.x * self.x + self.y * self.y + self.z * self.z).sqrt()
     }
 
-    pub fn squared_length(&self) -> f64 {
+    pub fn squared_length(&self) -> f32 {
         self.x * self.x + self.y * self.y + self.z * self.z
     }
 
-    pub fn dot(&self, v: Vec3) -> f64 {
+    pub fn dot(&self, v: Vec3) -> f32 {
         self.x * v.x + self.y * v.y + self.z * v.z
     }
 
@@ -73,16 +73,16 @@ impl ops::Neg for Vec3 {
     }
 }
 
-impl ops::Mul<f64> for Vec3 {
+impl ops::Mul<f32> for Vec3 {
     type Output = Vec3;
-    fn mul(self, t: f64) -> Vec3 {
+    fn mul(self, t: f32) -> Vec3 {
         Vec3::new(self.x * t, self.y * t, self.z * t)
     }
 }
 
-impl ops::Div<f64> for Vec3 {
+impl ops::Div<f32> for Vec3 {
     type Output = Vec3;
-    fn div(self, t: f64) -> Vec3 {
+    fn div(self, t: f32) -> Vec3 {
         Vec3::new(self.x / t, self.y / t, self.z / t)
     }
 }
@@ -111,14 +111,14 @@ impl ops::DivAssign<Vec3> for Vec3 {
     }
 }
 
-impl ops::MulAssign<f64> for Vec3 {
-    fn mul_assign(&mut self, t: f64) {
+impl ops::MulAssign<f32> for Vec3 {
+    fn mul_assign(&mut self, t: f32) {
         *self = Vec3::new(self.x * t, self.y * t, self.z * t);
     }
 }
 
-impl ops::DivAssign<f64> for Vec3 {
-    fn div_assign(&mut self, t: f64) {
+impl ops::DivAssign<f32> for Vec3 {
+    fn div_assign(&mut self, t: f32) {
         *self = Vec3::new(self.x / t, self.y / t, self.z / t);
     }
 }
